[Fix] 파일 올라가지 않는 문제 해결
