+++ AlarmButler/WorldClock/Views/WorldClockTVCell.swift
@@ -1,8 +1,8 @@
 //
-//  WorldClockViewCell.swift
+//  TableViewCell.swift
 //  AlarmButler
 //
-//  Created by t2023-m0099 on 2/7/24.
+//  Created by t2023-m0099 on 2/12/24.
 //
 
 import UIKit
@@ -18,16 +18,16 @@ class WorldClockViewCell: UITableViewCell {
         fatalError("init(coder:) has not been implemented")
     }
     
-
+    
     override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code
     }
-
+    
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
-
+        
         // Configure the view for the selected state
     }
-
+    
 }
